Revert "chore: move config related files to evmd/config"
